Add login flow and landing page
